wip: fix legacy climb cache invalidation
