+++ package.json
@@ -7,7 +7,7 @@
     "@algolia/autocomplete-js": "1.7.1",
     "@algolia/autocomplete-theme-classic": "1.7.1",
     "@apollo/client": "^3.10.1",
-    "@apollo/experimental-nextjs-app-support": "^0.10.0",
+    "@apollo/experimental-nextjs-app-support": "^0.11.9",
     "@dnd-kit/core": "^6.1.0",
     "@dnd-kit/sortable": "^8.0.0",
     "@dnd-kit/utilities": "^3.2.2",
