+++ src/js/hooks/useUpdateClimbsCmd.tsx
@@ -3,8 +3,8 @@ import { GraphQLError } from 'graphql'
 import { toast } from 'react-toastify'
 import { graphqlClient } from '../graphql/Client'
 import { MUTATION_UPDATE_CLIMBS, MUTATION_DELETE_CLIMBS, UpdateClimbsInput, DeleteManyClimbsInputType } from '../graphql/gql/contribs'
-import { refreshPage } from './useUpdateAreasCmd'
-import { invalidateAreaPageCache } from '../utils'
+import { invalidateAreaPageCache, invalidateClimbPageCache } from '../utils'
+import { legacyInvalidateClimbPageCache } from '../legacyInvalidateClimbPageCache'
 
 type UpdateClimbCmdType = (input: UpdateClimbsInput) => Promise<void>
 type DeleteClimbsCmdType = (idList: string[]) => Promise<number>
@@ -42,7 +42,8 @@ export default function useUpdateClimbsCmd ({ parentId, accessToken = '', onUpda
         const idList = Array.isArray(updateClimbs) ? updateClimbs : []
 
         idList.forEach(climbId => {
-          void refreshPage(`/api/revalidate?c=${climbId}`)
+          void legacyInvalidateClimbPageCache(climbId)
+          void invalidateClimbPageCache(climbId)
         })
 
         // Rebuild the parent area page
